Architecture improvements
