+++ src/components/rank-display/rank-display.styles.scss
@@ -1,5 +1,5 @@
 .rank-display {
-    $icon-dims: 50px;
+    $icon-dims: 4rem;
     display: flex;
     flex-direction: column;
     justify-content: center;
