+++ src/components/rank-info/rank-info.styles.scss
@@ -1,24 +1,30 @@
 .rank-info {
     display: grid;
-    gap: 15px;
     grid-template-columns: 2fr 1fr;
-    grid-template-rows: 2fr 1fr 2fr;
+    grid-template-rows: 1fr 1fr 1fr;
     grid-template-areas:
         "rank-img rank-name"
-        "rank-img rank-displays"
-        "rank-img tier";
-    .tier {
-        grid-area: tier;
-        background-color: red;
-        display: flex;
-        justify-content: center;
-        align-items: center;
+        "rank-img tier"
+        "rank-displays rank-displays";
+}
+@media (min-width: 767px) {
+    .rank-info {
+        gap: 15px;
+        grid-template-columns: 2fr 1fr;
+        grid-template-rows: 2fr 1fr 2fr;
+        grid-template-areas:
+            "rank-img rank-name"
+            "rank-img rank-displays"
+            "rank-img tier";
     }
-    .rank-name {
-        grid-area: rank-name;
-        background-color: rebeccapurple;
-        display: flex;
-        justify-content: center;
-        align-items: center;
+}
+@media (min-width: 1200px) {
+    .rank-info {
+        grid-template-columns: 1fr 1fr;
+        grid-template-rows: 1fr 1fr 60vh;
+        grid-template-areas:
+            "rank-name tier"
+            "rank-displays rank-displays"
+            "rank-img rank-img";
     }
 }
