+++ src/components/rank-name/rank-name.styles.scss
@@ -1,0 +1,7 @@
+.rank-name {
+    grid-area: rank-name;
+    background-color: rebeccapurple;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
