+++ src/components/rank-tier/rank-tier.styles.scss
@@ -1,0 +1,7 @@
+.tier {
+    grid-area: tier;
+    background-color: red;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
