+++ src/components/rank/rank.styles.scss
@@ -1,0 +1,6 @@
+.rank {
+    display: grid;
+    grid-template-columns: 1fr;
+    grid-template-rows: 1fr;
+    grid-template-areas: "rank-info";
+}
