+++ src/index.css
@@ -15,6 +15,7 @@ body {
         "Helvetica Neue", sans-serif;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
+    min-width: 300px;
 }
 
 code {
